reduce size of `struct syscall_cache_t`

+++ pkg/security/ebpf/c/syscalls.h
@@ -46,8 +46,8 @@ struct linux_binprm_t {
 struct syscall_cache_t {
     struct policy_t policy;
     u64 type;
-    u16 discarded;
-    u16 ad_saved;
+    u8 discarded;
+    u8 ad_saved;
     u8 async;
 
     struct dentry_resolver_input_t resolver;
